refactor: Move Snapshot into model package
