+++ packages/cohort/src/model/mod.rs
@@ -1,3 +1,4 @@
 pub mod bank_account;
 pub(crate) mod json;
+pub mod snapshot;
 pub mod talos_state;
